Added results passing and change DICOM directory to correspond with reconstruction name
Added returning of the reconstructed image as an option.
Also fixed the naming of the DICOM directory of the reconstructed image to correspond with the reconstruction name.
Also changed resolving the archive directory path based in the patient ID in the header, rather than parsing subdirectory name, as it didn't work well with baseline reconstruction during simulation.